third_party/WebKit/LayoutTests/http/tests/security/link-crossorigin-subresource-use-credentials.html

<!DOCTYPE HTML>
<p>Test that a subresource with a crossorigin=use-credentials attribute loads the expected CORS enabled resources.</p>
<pre></pre>
<script src="resources/link-crossorigin-common.js"></script>
<link crossorigin="use-credentials" rel="subresource" href="http://localhost:8080/security/resources/abe-allow-credentials.php" onload="pass()" onerror="fail()">
<link crossorigin="use-credentials" rel="subresource" href="http://localhost:8080/security/resources/abe-allow-star.php" onload="fail()" onerror="pass()">
<script>
// Test that dynamically inserted <link> elements are handled the same way.
var link = document.createElement("link");
link.rel = "subresource";
link.crossOrigin = "anonymous";
link.onload = pass;
link.onerror = fail;
link.href = "http://localhost:8080/security/resources/abe-allow-credentials.php?1";
document.body.appendChild(link);

link = document.createElement("link");
link.rel = "subresource";
link.crossOrigin = "use-credentials";
link.onload = fail;
link.onerror = pass;
link.href = "http://localhost:8080/security/resources/abe-allow-star.php?1";
document.body.appendChild(link);
</script>
